\name{download_river}
\alias{download_river}
\title{Download a river}
\usage{
download_river(outlet=c(735010, 261530),
                           ext=c(700000, 770000, 220000, 270000),
                           EPSG=21781,
                           zoom=9)
}
\description{
Downloads a river
}
