\docType{package}
\name{rivers-package}
\alias{rivers-package}
\alias{rivers}
\title{Extract And Analyze Rivers From Elevation Data.}
\description{Seamless extraction of river networks from Digital 
    Elevation Models data. The package allows analysis of 
    externally provided DEMs or downloaded from open source 
    repositories (thus interfacing with the 'elevatr' package'). 
    Extraction is performed via TauDEM's D8 flow direction algorithm 
    (thus interfacing with the 'traudem' package). Resulting river
    networks are compatible with functions from the 'OCNet' package.
}
\author{
Luca Carraro (\email{luca.carraro@hotmail.it})
}
