\name{river-class}

\docType{class}

\alias{river}
\alias{river-class}

\alias{show,river-method}
\alias{$,river-method}
\alias{$<-,river-method}
\alias{names,river-method}


\title{river class}

\description{
A \code{river} object contains information on river attributes at different aggregation levels. It can represent a real river network 
(obtained via \code{\link{extract_river}}) or an optimal channel network (obtained via \code{\link{create_OCN}}).

The content of a \code{river} object can be treated as a list. For information on the aggregation levels and on the content of a 
\code{river} object, see \code{\link{OCNet-package}}.
}
